refactor splash screen generation logic
